Add support for server presets.
To make it even easier for non-technical people to get started with IRC,
this change adds support for presets defined in config.js.  In such a
deployment, instead of requiring users to type in all the server
configuration parameters, subway automatically connects to the preset
server and joins a predefined set of channels.  After that, users may
join and leave channels as normal, but they cannot connect to other
servers.