Fixes bug where one could register with empty username or password.
After commit 48bf2d64d4db it became possible to register with an empty username
and/or password. In fact, the checks ensuring a value in the username and
password fields were referencing legacy classes and no longer working. I thus
changed to a horizontal form per Twitter Bootstrap documentation[0] and added
some validation states informing the user of what is wrong.

[0]: http://getbootstrap.com/2.3.2/base-css.html#forms